AppCode fixes

+++ SUS Inspector/RSVerticallyCenteredTextFieldCell/RSVerticallyCenteredTextFieldCell.m
@@ -26,7 +26,7 @@
 		NSSize textSize = [self cellSizeForBounds:theRect];
 
 		// Center that in the proposed rect
-		float heightDelta = newRect.size.height - textSize.height;	
+		CGFloat heightDelta = newRect.size.height - textSize.height;
 		if (heightDelta > 0)
 		{
 			newRect.size.height -= heightDelta;
